% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_ss_versionNumeric}
\alias{check_ss_versionNumeric}
\title{Check Stock Synthesis Numeric Version Number}
\usage{
check_ss_versionNumeric(x, min_version = 3.3)
}
\arguments{
\item{x}{Numeric integer representing Stock Synthesis Version, found in
ss_versionNumeric}

\item{min_version}{Minimum SS version to compare Stock Synthesis object
files. By default, it is 3.3.}
}
\description{
Custom checkmate function to compare Stock Synthesis Numeric Version Number.
Includes minimum version check.
}
