% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_basemodel.R
\name{copy_n_boot_sso}
\alias{copy_n_boot_sso}
\title{Copy Bootstrap Run Output Files to Bootstrap Directory}
\usage{
copy_n_boot_sso(
  boot_dir,
  n_boot,
  copy_compReport = TRUE,
  copy_covar = TRUE,
  copy_warning = TRUE
)
}
\arguments{
\item{boot_dir}{Path where the bootstrap files will be saved. Bootstrap runs}

\item{n_boot}{Number of Bootstraps}

\item{copy_compReport}{Copy compReport Bootstrap Run to Bootstrap
Directory? Default is TRUE}

\item{copy_covar}{Copy covar Bootstrap Run to Bootstrap Directory? Default
is TRUE}

\item{copy_warning}{Copy warning Bootstrap Run to Bootstrap Directory?
Default is TRUE.}
}
\description{
Copies the Stock Synthesis Report File, Composition Data, Cumulative
Summaries, and Log of Warnings from each Bootstrap run to the main
bootstrap directory. Each copied bootstrap run file will be renamed to
have a bootstrap run identifier.
}
\keyword{Bootstrap}
\keyword{internal}
