% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{default_cv_fleets_process_error}
\alias{default_cv_fleets_process_error}
\title{Default Coefficient of Variation for process error parameters with fleets}
\usage{
default_cv_fleets_process_error(max_age, num_fleets, value = 0.1)
}
\arguments{
\item{max_age}{Max Age. Determines the length of vector.}

\item{num_fleets}{Number of Fleets. For}

\item{value}{Default value for CV.}
}
\description{
Returns a matrix containing a default Coefficient of Variation (CV) value.
The dimensions of the matrix is determined by max_age over the number of
fleets.
}
\examples{
\dontrun{
  get_cv_fleets_process_error(10, 5, 0.1)
}


}
