% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{default_cv_process_error}
\alias{default_cv_process_error}
\title{Process Error parameter's default Coefficient of Variation}
\usage{
default_cv_process_error(max_age, value = 0.1)
}
\arguments{
\item{max_age}{Max Age. Determines the length of vector.}

\item{value}{Default value for CV.}
}
\description{
Returns the vector of Coefficient of Variation (CV) values. The length of
the vector is determined by max_age.
}
\examples{
\dontrun{
  get_cv_process_error(10)
}

}
