% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{export_ss_objectlist_year}
\alias{export_ss_objectlist_year}
\title{Export Stock Synthesis Object List for AGEPRO by Year}
\usage{
export_ss_objectlist_year(ss_objectlist, ss_agepro)
}
\arguments{
\item{ss_objectlist}{Stock Synthesis list object data, primary from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{ss_agepro}{Stock Synthesis Object List for AGEPRO parameters}
}
\description{
Exports the Stock Synthesis Object List parameters that AGEPRO uses, using
the Year time series.
}
\details{
The function will gather the parameter and CV table for Maturity, Fishery
Selectivity Of Age, Weights of Age (Jan-1, Spawning Stock
Biomass, Mid-Year), and Catch at Age. Each years \code{CatchByFleet} is used
to calculate the proportion of total catch.

For Maturity, maturity if age, starting at age 1. The length at age then
use maturity to give to calculate maturity at age
\deqn{P_{mature}(L) = \frac{1}{(1 + exp(beta*(L-L_{50})))}}
}
