% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract_end_year}
\alias{extract_end_year}
\title{Extracts end year}
\usage{
extract_end_year(ss_objectlist)
}
\arguments{
\item{ss_objectlist}{Stock Synthesis list object data, primary from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}
}
\description{
Filters out the end year value based on the stock synthesis model
}
