% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_bootstrap.R
\name{format_duration_string}
\alias{format_duration_string}
\title{Formatted time duration string between two time periods}
\usage{
format_duration_string(start, end)
}
\arguments{
\item{start}{POSIXct value indicating the "start time". Value must be less
than end or "end time."}

\item{end}{POSIX value indicating "end time". Value must be greater than
start or "start time"}
}
\description{
Helper function that returns the time difference of two POSIXct values
formatted as \verb{\%dh \%dm \%.1fs}.
}
\keyword{internal}
