% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{get_param_by_fleet}
\alias{get_param_by_fleet}
\title{Get Parameter Value per fleet}
\usage{
get_param_by_fleet(
  ss_objectlist,
  colname_prefix,
  timestep = c("Year", "Quarter")
)
}
\arguments{
\item{ss_objectlist}{Stock Synthesis list object data, primary from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{colname_prefix}{Character string to select to target parameter with for each fleet}

\item{timestep}{"Year" or "Quarter": Indicates is you are running AGEPRO
with a yearly time step or as quarters as years. "Year" as Default.}
}
\description{
Get Parameter Value per fleet
}
