% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{get_ss_objectlist_parameter}
\alias{get_ss_objectlist_parameter}
\title{Get a parameter and its values from the Stock Synthesis Object List}
\usage{
get_ss_objectlist_parameter(ss_objectlist, ss_label)
}
\arguments{
\item{ss_objectlist}{Stock Synthesis list object data, primary from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{ss_label}{Name of the ss_objectlist parameter to extract values from.}
}
\description{
Convenience function to get parameter values from the Stock Synthesis Object
List.
}
