% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{get_WAA_growth}
\alias{get_WAA_growth}
\title{Gets growth information for weight of Age parameter values.}
\usage{
get_WAA_growth(ss_objectlist, timestep = c("Year", "Quarter"))
}
\arguments{
\item{ss_objectlist}{Stock Synthesis list object data, primary from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{timestep}{"Year" or "Quarter": Indicates is you are running AGEPRO
with a yearly time step or as quarters as years. "Year" as Default.}
}
\description{
Convenience function to get Growth Information  from
the Stock Synthesis Object List to calculate Weight of Age. Method will
vary depending on timestep used.
}
