% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_bootstrap.R
\name{nboot_setup}
\alias{nboot_setup}
\title{Setting up a Bootstrapping Directory with a Base Stock Synthesis Model}
\usage{
nboot_setup(basemodel_dir, out_dir, n_boot, ss3_exe = "ss3.exe")
}
\arguments{
\item{basemodel_dir}{basemodel directory}

\item{out_dir}{Target output directory}

\item{n_boot}{Number of Bootstraps}

\item{ss3_exe}{ss3_exe}
}
\value{
Returns a list contain the location of the Bootstrap
Subdirectories.
}
\description{
Creates a bootstrapped model based on the base components of the Stock
Synthesis Model (Control File, Starter File, Forecast File, Stock
Synthesis Parameters), that is dependent to the number of bootstraps is
needed.
}
\details{
Bootstrap sub directories are created in the target out_dir. Each bootstrap
subdirectory will have a copy of the basemodel Control File, Forecast File,
and Stock Synthesis Parameters. Each Bootstrap run starter file will be
saved as Data Boot Run File. The bootstrap Starter field  \code{datfile} will
point to a \verb{data_boot_<n_boot>.ss} file as the starter file; This will
indicate Stock Synthesis to start at that bootstrap.
}
\keyword{Bootstrap}
