% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{not_implmented_ageproR}
\alias{not_implmented_ageproR}
\title{ageproR Not Implemented Error}
\usage{
not_implmented_ageproR()
}
\description{
Returns error to indicate feature or action isn't currently implemented
with ageproR.
}
\keyword{Internal}
\keyword{messages}
