% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_bootstrap.R
\name{run_nboot_setup}
\alias{run_nboot_setup}
\title{Function Wrapper to Setup Bootstrapped runs in parallel via futures}
\usage{
run_nboot_setup(basemodel_dir, out_dir, n_boot, ss3_exe = "ss3.exe")
}
\arguments{
\item{basemodel_dir}{basemodel directory}

\item{out_dir}{Target output directory}

\item{n_boot}{Number of bootstraps}

\item{ss3_exe}{ss3_exe}
}
\description{
This is a function wrapper to run the function nboot_setup via future
}
\details{
#' @returns Returns a list contain the location of the Bootstrap
Subdirectories.
}
\keyword{Bootstrap}
