% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_bootstrap.R
\name{run_r4ss_with_spinner}
\alias{run_r4ss_with_spinner}
\title{Run r4ss with a spinner on Rconsole}
\usage{
run_r4ss_with_spinner(out_dir, ss3_exe)
}
\arguments{
\item{out_dir}{Ouptut directory where the stock synthesis files will be saved.}

\item{ss3_exe}{Executable name. Can be just the name of the executable
file if it is in the specified directory or in the user's PATH. Can also
include the absolute path or a path relative to the specified directory.
Needs to be a single character string, not a vector. On Windows, exe can
optionally have the .exe extension appended; on Unix-based systems
(i.e., Mac and Linux), no extension should be included.}
}
\description{
Allows to display a splnner ojection on the R Console to indicate running
Stock Synthesis process This is helped by running the r4ss process in \
another R session.
}
\keyword{internal}
