% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp_model_recruitment.R
\name{set_inp_model_recruit_data}
\alias{set_inp_model_recruit_data}
\title{Use Stock Synthesis Object data to create AGEPRO model recruit data.}
\usage{
set_inp_model_recruit_data(inp_model, ss_agepro)
}
\arguments{
\item{inp_model}{AGEPRO model}

\item{ss_agepro}{Stock Synthesis Object List for AGEPRO parameters}
}
\description{
Helper function to set AGEPRO model recruitment model data.
}
