% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp_model_recruitment.R
\name{set_inp_model_recruit_prob}
\alias{set_inp_model_recruit_prob}
\title{Sets up AGEPRO Recruitment Probability values}
\usage{
set_inp_model_recruit_prob(inp_model, recruit_model_prob)
}
\arguments{
\item{inp_model}{AGEPRO model R6class from agerpoR}

\item{recruit_model_prob}{The Recruitment Probabilities for each recruitment
model.}
}
\description{
Convenience function to set the recruitment probability of a agepro model.
}
