% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{set_ss_agepro_WAA}
\alias{set_ss_agepro_WAA}
\title{Sets weight of Age parameter values based from the Stock Synthesis Object List}
\usage{
set_ss_agepro_WAA(ss_agepro, WAA_param, ss_objectlist, ss_label)
}
\arguments{
\item{ss_agepro}{Stock Synthesis Object List for AGEPRO parameters}

\item{WAA_param}{Weight of Age Parameter}

\item{ss_objectlist}{Stock Synthesis list object data, primary from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{ss_label}{Name of the ss_objectlist parameter to extract values from}
}
\description{
Convenience function to get Weight of Age parameter and CV values from
the Stock Synthesis Object List to the AGEPRO object list.
}
