% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_agepro_inpfile.R
\name{setup_agepro_inpfile}
\alias{setup_agepro_inpfile}
\title{AGEPRO Model class creations from ss_agepro objectlist}
\usage{
setup_agepro_inpfile(
  ss_agepro,
  num_years,
  num_pop_sims,
  recruit_models,
  recruit_model_prob,
  bsn_file,
  num_boot = 100,
  model_name = "untitled"
)
}
\arguments{
\item{ss_agepro}{Stock Synthesis Object List for AGEPRO parameters}

\item{num_years}{Number of Years}

\item{num_pop_sims}{Number of population simulations}

\item{recruit_models}{Vector of recruitment model numbers}

\item{recruit_model_prob}{Recruitment proabbility}

\item{bsn_file}{Bootstrap file}

\item{num_boot}{Number of bootstraps}

\item{model_name}{Model name}
}
\description{
Conveince function to setup AGEPRO model class with AGEPRO input model
file functions.
}
\examples{
\dontrun{


#Continuing from ss_output_export_agepro example
bsn_file <- file.path(bootstrap_dir,"bootstrap.bsn")

inp_file <-
  setup_agepro_inpfile(ss_agepro,
                      num_years = 8,
                      num_pop_sims = 1000,
                      recruit_models = c(5,3),
                      recruit_model_prob = c(0.6, 0.4),
                      bsn_file)

}


}
