% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_basemodel.R
\name{setup_bootstrap_dir}
\alias{setup_bootstrap_dir}
\title{Create SS data for bootstrap.}
\usage{
setup_bootstrap_dir(basemodel_dir, boot_dir, n_boot = 1, ss3_exe = "ss3.exe")
}
\arguments{
\item{basemodel_dir}{Target base model directory path}

\item{boot_dir}{Target path for bootstrap output}

\item{n_boot}{number of bootstraps}

\item{ss3_exe}{ss3_exe}
}
\description{
Embedded Helper function to copy base model stock synthesis data for each
bootstrap and runs.
}
\keyword{internal}
