% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_basemodel.R
\name{setup_ss_basemodel}
\alias{setup_ss_basemodel}
\title{Creates a Stock Synthesis base model and bootstrap runs}
\usage{
setup_ss_basemodel(
  basemodel_dir,
  ss3_path,
  n_boot = 100,
  seed = NULL,
  ss3_exe = "ss3",
  bootstrap_outdir = file.path(this.path::this.proj())
)
}
\arguments{
\item{basemodel_dir}{where the base model SS files are located}

\item{ss3_path}{File path of the Stock Synthesis Binary}

\item{n_boot}{Number of Bootstraps}

\item{seed}{Pusedorandom Number Seed}

\item{ss3_exe}{Executable name. Can be just the name of the executable file if it is in the specified directory or in the user's PATH. Can also include the absolute path or a path relative to the specified directory. Needs to be a single character string, not a vector. On Windows, exe can optionally have the .exe extension appended; on Unix-based systems (i.e., Mac and Linux), no extension should be included.}

\item{bootstrap_outdir}{Path where the bootstrap file and bootstrap runs will be saved}
}
\description{
Sets up a Stock Synthesis Base model including individual bootstrap runs.
Function includes helper functions to to help with running SS in parallel
}
