% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_bootstrap.R
\name{ss_model_n_boot}
\alias{ss_model_n_boot}
\title{Create Bootstrap runs}
\usage{
ss_model_n_boot(basemodel_dir, boot_dir, n_boot, ss3_exe = "ss3.exe")
}
\arguments{
\item{basemodel_dir}{basemodel directory}

\item{boot_dir}{boostrap directory}

\item{n_boot}{number of bootstraps}

\item{ss3_exe}{ss3_exe}
}
\description{
Set up each bootstrap run in its own folder, to help with running SS in
parallel. The Base Model Stock Synthesis Files are copied to each
Bootstrap run.
}
\details{
Each Bootstrap run starter file will be saved as Data Boot Run File.
}
\keyword{Bootstrap}
\keyword{internal}
