% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_output_export_agepro.R
\name{ss_output_export_agepro}
\alias{ss_output_export_agepro}
\title{Export Stock Synthesis Model data for AGEPRO Input Files}
\usage{
ss_output_export_agepro(ss_objectlist, timestep = c("Year", "Quarter"))
}
\arguments{
\item{ss_objectlist}{Stock Synthesis list object data, primary from
\code{\link[r4ss:SS_output]{r4ss::SS_output()}}}

\item{timestep}{"Year" or "Quarter": Indicates is you are running AGEPRO
with a yearly time step or as quarters as years. "Year" as Default.}
}
\value{
This fuction Returns a list containing the the Stock Synthesis following values:
\itemize{
\item \code{Nfleets} is the total number of catch fleets in the SS3 Report file
\item \code{alpha} is the alpha parameter of a Beverton Holt or Richards stock assessment function
\item \code{beta} is the beta parameter of a beverton holt or richards stock assessment function
\item \code{BH_Var} the sigmaR from the SS3 model, or the variance of the recruitment deviations
\item \code{RecruitmentObs} is a datatable containing the predicted recruitment, SSB, and recruitment deviation for each year or quarter
\item \code{MaxAge} is the Amax parameter from the SS3 model or maximum age, note for the Quarter time step this is maximum age in quarters (or \verb{MaxAge\\*4})
\item \code{MaturityAtAge} is a vector containing the probability of maturity at age by year or quarter
\item \code{MaturityAtAgeCV} is a vector of the CV for the probability of maturity at age, note this (and all CVs) are set to 0.1 as a default but can be adjusted after running the script
\item \code{Fishery_SelAtAge} datatables of the selectivity by age and CV for each of the Nfleets
\item \code{Fishery_SelAtAgeCV} datatables of the selectivity by age and CV for each of the Nfleets
\item \code{NatMort_atAge} vectors of the natural mortality by age in years or quarters and their CV
\item \code{NatMort_atAgeCV} vectors of the natural mortality by age in years or quarters and their CV
\item \code{Jan_WAA} is the weight-at-age of the stock on Jan-1
\item \code{Jan_WAACV} CV of Weight-Of-Age of stock on Jan-1
\item \code{MidYear_WAA} is the weight-at-age of the stock on July-1
\item \code{MidYear_WAACV} CV of weight-at-age of the stock on July-1
\item \code{SSB_WAA} SSB Weight of Age
\item \code{SSB_WAACV} SSB Weight-of-Age
\item \code{SSB_WAACV} CV of SSB Weight-Of-Age
\item \code{CatchAtAge} Catch Weight-of-age by Nfleet
\item \code{CatchAtAgeCV} Catch Weight-of-Age CV by Nfleet
\item \code{CatchByFleet} is the total catch by fleet in the last year/quarter of the model
\item \code{FByFleet} Fishing Mortality by Fleet
}
}
\description{
Function to extract input data needed from the SS model to use in an agepro projection model
}
\details{
Units: SSB and catch are in metric tons of biomass, recruitment is in 1000s of fish,
catch/weight at ages are all in kilograms, alpha and beta parameters are converted from metric tons/1000s
of fish to kilograms and numbers of fish by multiplying by 1000
}
\examples{
\dontrun{

 # Setup For Bootstrap File and Directory
 basemodel_dir <- file.path(find.package("sso.agepro"),"01_base")
 bootstrap_dir <- file.path(tempdir())
 setup_ss_bootstrap(basemodel_dir, bootstrap_dir, n_boot = 10)

 # Call r4ss::SS_output to extract from Report.sso
 ss_objectlist <- r4ss::SS_output(dir = bootstrap_dir, verbose = FALSE, printstats = FALSE)

 ss_agepro <- ss_output_export_agepro(ss_objectlist)

 #Filter For unique selectivity

}

}
\author{
Michelle Sculley

Eric Fletcher
}
