% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{subset_empirical_recobs}
\alias{subset_empirical_recobs}
\title{Subset empirical recruitment observation table}
\usage{
subset_empirical_recobs(ss_agepro, start_yr, end_yr)
}
\arguments{
\item{ss_agepro}{Stock Synthesis Object List for AGEPRO parameters}

\item{start_yr}{First year of time period to filter "Yr" field}

\item{end_yr}{Last Year of time period to filter "Yr" field}
}
\description{
Helper method to extract the empirical recruitment observation table from
the Stock synthesis report objectlist.
}
