% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_r4ss_parallel.R
\name{suppress_r4ss_messages}
\alias{suppress_r4ss_messages}
\title{Suppress r4ss verbose messages}
\usage{
suppress_r4ss_messages(x)
}
\arguments{
\item{x}{Function Call}
}
\description{
Helper Function to suppress the verbose messages the "chatty" r4ss during
the run_parallel function.
}
