% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unique_selectivity_fleets}
\alias{unique_selectivity_fleets}
\title{Fleets with a unique length based selectivity}
\usage{
unique_selectivity_fleets(ss_objectlist)
}
\arguments{
\item{ss_objectlist}{Stock Synthesis list object data, primary from \code{\link[r4ss:SS_output]{r4ss::SS_output()}}}
}
\description{
(Review) For building AGEPRO input files, from Stock Synthesis length based
selectivity data.
}
\examples{
\dontrun{

basemodel_dir <- file.path(find.package("sso.agepro"),"01_base")
base_model <- r4ss::SS_output(basemodel_dir)

# Returns Catch Fleet Numbers matching unique selectvity criteria.
unique_selectivity_fleets(base_model)

# Number of Fleets
length(unique_selectivity_fleets(base_model))

}


}
