% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unsupported_model}
\alias{unsupported_model}
\title{Unsupported model error}
\usage{
unsupported_model()
}
\description{
Returns error to indicate unsupported recruitment.
}
\keyword{Internal}
\keyword{messages}
