% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_bootstrap.R
\name{endyr_model}
\alias{endyr_model}
\title{Extracts end year of model}
\usage{
endyr_model(boot_dir)
}
\arguments{
\item{boot_dir}{Path where the bootstrap files will be saved. Bootstrap runs}
}
\description{
Embedded Helper function to return end year of model (endyr) and derived
quants
}
\keyword{Bootstrap}
\keyword{internal}
