% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{set_empirical_distribution_data}
\alias{set_empirical_distribution_data}
\title{Use Stock Synthesis data to Set Empirical Recruitment Distribution.}
\usage{
set_empirical_distribution_data(inp_model, ss_agepro, irec = 1)
}
\arguments{
\item{inp_model}{AGEPRO model}

\item{ss_agepro}{Stock Synthesis Object List for AGEPRO parameters}

\item{irec}{Index of (Multi) recruitment Model}
}
\description{
Helper function to set Empirical Recruitment Distribution data (Model #3)
from stock synthesis report objectlist data.
}
