% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_agepro_inpfile.R
\name{setup_agepro_inpfile}
\alias{setup_agepro_inpfile}
\title{AGEPRO Model class creations from ss_agepro objectlist}
\usage{
setup_agepro_inpfile(
  ss_agepro,
  num_years,
  num_pop_sims,
  recruit_models,
  recruit_model_prob,
  model_name = "untitled"
)
}
\arguments{
\item{ss_agepro}{Stock Synthesis Object List for AGEPRO parameters}

\item{num_years}{Number of Years}

\item{num_pop_sims}{Number of population simulations}

\item{recruit_models}{Vector of recruitment model numbers}

\item{recruit_model_prob}{Recruitment proabbility}

\item{model_name}{Model name}
}
\description{
Conveince function to setup AGEPRO model class with AGEPRO input model
file functions.
}
