% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_basemodel.R
\name{setup_ss_basemodel}
\alias{setup_ss_basemodel}
\title{Creates a Stock Synthesis base model and bootstrap runs}
\usage{
setup_ss_basemodel(
  basemodel_dir,
  bootstrap_outdir,
  n_boot = 100,
  seed = 123,
  ss3_exe = "ss3"
)
}
\arguments{
\item{basemodel_dir}{File path where the Base Model Stock Synthesis files
are located}

\item{bootstrap_outdir}{Path where the bootstrap file and bootstrap runs will be saved}

\item{n_boot}{Number of Bootstraps}

\item{seed}{Pusedorandom Number Seed. Default is 123}

\item{ss3_exe}{Executable name. Can be just the name of the executable
file if it is in the specified directory or in the user's PATH. Can also
include the absolute path or a path relative to the specified directory.
Needs to be a single character string, not a vector. On Windows, exe can
optionally have the .exe extension appended; on Unix-based systems
(i.e., Mac and Linux), no extension should be included.}
}
\description{
Sets up a Stock Synthesis Base model including individual bootstrap runs.
Function includes helper functions to to help with running SS in parallel
}
\examples{
\dontrun{
#Import Basemodel from inst/01_base and output Bootstraps to inst/bsn_output
basemodel_dir <- file.path(find.package("sso.agepro"),"inst","01_base")
output_dir <- file.path(tempdir(),"inst","bsn_output")

setup_ss_basemodel(basemodel_dir, bootstrap_output = output_dir, n_boot = 10)

}


}
\keyword{Bootstrap}
