% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_basemodel.R
\name{write_bsn_file}
\alias{write_bsn_file}
\title{Write an Age Based Bootstrap File}
\usage{
write_bsn_file(endyr, output_dir, n_boot = 1)
}
\arguments{
\item{endyr}{End of Year Model}

\item{output_dir}{Target path to write AGEPRO bootstrap file}

\item{n_boot}{number of Bootstraps}
}
\description{
Writes an Age Based Bootstrap File
}
\keyword{internal}
