% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ss_basemodel.R
\name{write_bsn_file}
\alias{write_bsn_file}
\title{Write an Age Based Bootstrap File}
\usage{
write_bsn_file(boot_dir, endyr, n_boot = 1)
}
\arguments{
\item{boot_dir}{Path where the bootstrap files will be saved. Bootstrap runs}

\item{endyr}{End of Year Model}

\item{n_boot}{Number of Bootstraps}
}
\description{
Writes an Age Based Bootstrap File. Output is written to the bootstrap
directory.
}
\keyword{Bootstrap}
\keyword{internal}
